Fix bug in input function and Add math functionality
Fixed a bug that allowed use to add or removed negative sign to zero
and when the decimal is only applied on numbers.
Added functionality that allows user to make math operations such as
adding and subtracting. Also added ability to do many math operations
without having to press the = button